Remove old stuff, disable verbose mode, configure coverage as previously
