--- .c8rc.json
@@ -1,12 +1,0 @@
-{
-  "all": true,
-  "include": [
-    "bin",
-    "lib"
-  ],
-  "reporter": [
-    "html",
-    "lcov",
-    "text"
-  ]
-}
