+++ jest.config.js
@@ -3,25 +3,25 @@
 module.exports = {
     clearMocks: true,
     resetMocks: true,
-    coveragePathIgnorePatterns: [
-        '/node_modules/'
+    coverageProvider: 'v8',
+    coverageReporters: ['html', 'lcov', 'text'],
+    collectCoverageFrom: [
+        'bin/*.js',
+        'lib/**/*.js',
+        '!**/node_modules/**'
     ],
-
     moduleFileExtensions: [
         'js',
         'json'
     ],
-
     testMatch: [
-        '**/*.test.jest.js'
+        '**/*.test.js'
     ],
-
     testPathIgnorePatterns: [
         '/node_modules/'
     ],
-
     setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
     globalSetup: '<rootDir>/jest.setup.global.js',
-    verbose: true,
+    verbose: false,
     testTimeout: 10_000
 };
