+++ test/shape.test.js
@@ -1,13 +1,13 @@
 'use strict';
 
-const assert = require('assert').strict;
-const { readFileSync } = require('fs');
 const path = require('path');
+const { Buffer } = require('buffer');
 const { DOMParser } = require('@xmldom/xmldom');
 const SVGSpriter = require('../lib/svg-sprite.js');
-const { isObject } = require('../lib/svg-sprite/utils/index.js');
+const calculateSvgDimensions = require('../lib/svg-sprite/utils/calculate-svg-dimensions.js');
+const { DEFAULT_XML_DECLARATION } = require('../lib/svg-sprite/constants.js');
 
-const { paths } = require('./helpers/constants.js');
+jest.mock('../lib/svg-sprite/utils/calculate-svg-dimensions.js');
 
 const expectations = [{
     svg: '46x46.svg',
@@ -23,45 +23,48 @@ const expectations = [{
     }
 }];
 
-describe('shape', () => {
-    expectations.forEach(expectation => {
-        it(`should calculate the dimensions if the ${expectation.svg} does not contain viewBox or height/width properties (${expectation.result.width}x${expectation.result.height})`, done => {
-            const spriter = new SVGSpriter({
-                shape: {
-                    dest: 'svg',
-                    dimension: {
-                        maxWidth: 4000,
-                        maxHeight: 4000
-                    }
+const TEST_SVG = '<svg xmlns="http://www.w3.org/2000/svg"></svg>';
+
+describe('testing shapes', () => {
+    it.each`
+            svg                    | dimension
+            ${expectations[0].svg} | ${expectations[0].result}
+            ${expectations[1].svg} | ${expectations[1].result}
+        `('should call calculateSvgDimensions if the $svg does not contain viewBox or height/width properties ($dimension)', async({
+        svg, dimension
+    }) => {
+        expect.hasAssertions();
+
+        const spriter = new SVGSpriter({
+            shape: {
+                dest: 'svg',
+                dimension: {
+                    maxWidth: 4000,
+                    maxHeight: 4000
                 }
-            });
-
-            const svgFilePath = path.join(paths.fixtures, `svg/special/without-dims/${expectation.svg}`);
-
-            spriter.add(
-                svgFilePath,
-                expectation.svg,
-                readFileSync(svgFilePath, 'utf-8')
-            );
-
-            spriter.compile((error, result) => {
-                try {
-                    assert.ifError(error);
-                    assert.equal(isObject(result), true);
-                    assert.notEqual(typeof result.shapes, 'undefined');
-                    assert.equal(Array.isArray(result.shapes), true);
-
-                    const svg = result.shapes[0]._contents.toString();
-                    const dom = new DOMParser().parseFromString(svg, 'text/xml');
-
-                    assert.equal(dom.documentElement.getAttribute('height'), expectation.result.height.toString());
-                    assert.equal(dom.documentElement.getAttribute('width'), expectation.result.width.toString());
-
-                    done();
-                } catch (compilationError) {
-                    done(compilationError);
-                }
-            });
+            }
         });
+
+        calculateSvgDimensions.mockReturnValueOnce(dimension);
+
+        const svgFilePath = path.join(__dirname, `fixture/svg/special/without-dims/${svg}`);
+
+        spriter.add(
+            svgFilePath,
+            svg,
+            Buffer.from(TEST_SVG)
+        );
+
+        expect(calculateSvgDimensions).toHaveBeenCalledWith(new DOMParser().parseFromString(`${DEFAULT_XML_DECLARATION}${TEST_SVG}`).toString());
+
+        const { result } = await spriter.compileAsync();
+
+        expect(result).toBeInstanceOf(Object);
+        expect(result.shapes).toBeInstanceOf(Array);
+
+        const dom = new DOMParser().parseFromString(result.shapes[0]._contents.toString(), 'text/xml');
+
+        expect(dom.documentElement.getAttribute('height')).toStrictEqual(dimension.height.toString());
+        expect(dom.documentElement.getAttribute('width')).toStrictEqual(dimension.width.toString());
     });
 });
