+++ test/svg-sprite/alignment/mixed.test.js
@@ -1,19 +1,14 @@
 'use strict';
 
-/* eslint-disable no-unused-expressions, max-nested-callbacks */
 const path = require('path');
 const fs = require('fs');
 const mustache = require('mustache');
-const should = require('should');
-const looksSame = require('looks-same');
 const sass = require('sass');
 const glob = require('glob');
 const SVGSpriter = require('../../../lib/svg-sprite.js');
 const { addFixtureFiles } = require('../../helpers/add-files.js');
 const writeFiles = require('../../helpers/write-files.js');
 const writeFile = require('../../helpers/write-file.js');
-const capturePuppeteer = require('../../helpers/capture-puppeteer.js');
-const compareSvg2Png = require('../../helpers/compare-svg-2-png.js');
 const removeTmpPath = require('../../helpers/remove-temp-path.js');
 const { paths } = require('../../helpers/constants.js');
 
@@ -21,17 +16,19 @@ const cwdAlign = path.join(paths.fixtures, 'svg/css');
 const align = glob.sync('**/*.svg', { cwd: cwdAlign });
 const previewTemplate = fs.readFileSync(path.join(__dirname, '../../tmpl/css.html'), 'utf-8');
 
+const tmpPath = path.join(paths.tmp, 'mixed');
+
 describe(`svg-sprite: with mixed alignment and ${align.length} SVG files`, () => {
+    beforeAll(removeTmpPath.bind(null, tmpPath));
+
     describe('with «view» mode, vertical layout and CSS render type', () => {
         let spriter = null;
         let data = null;
         let svgPath = null;
 
-        before(removeTmpPath);
-
-        before(done => {
+        beforeAll(async() => {
             spriter = new SVGSpriter({
-                dest: paths.tmp,
+                dest: tmpPath,
                 shape: {
                     align: path.join(paths.fixtures, 'yaml/align.mixed.yaml'),
                     dimension: {
@@ -41,7 +38,7 @@ describe(`svg-sprite: with mixed alignment and ${align.length} SVG files`, () =>
                 }
             });
             addFixtureFiles(spriter, align, cwdAlign);
-            spriter.compile({
+            const { result, data: cssData } = await spriter.compileAsync({
                 view: {
                     sprite: 'svg/view.vertical.mixed.svg',
                     layout: 'vertical',
@@ -52,42 +49,30 @@ describe(`svg-sprite: with mixed alignment and ${align.length} SVG files`, () =>
                         }
                     }
                 }
-            }, (error, result, cssData) => {
-                writeFiles(result);
-                data = cssData.view;
-                svgPath = path.basename(result.view.sprite.path);
-                done();
             });
+            writeFiles(result);
+            data = cssData.view;
+            svgPath = path.basename(result.view.sprite.path);
         });
 
-        it('creates visually correct sprite', done => {
-            compareSvg2Png(
-                path.join(paths.tmp, 'view/svg', svgPath),
-                path.join(paths.tmp, 'view/png/css.vertical.mixed.png'),
-                path.join(paths.expectations, '/png/css.vertical.mixed.png'),
-                path.join(paths.tmp, 'view/png/css.vertical.mixed.diff.png'),
-                done,
-                'The vertical sprite doesn\'t match the expected one!'
+        it('creates visually correct sprite', async() => {
+            expect.hasAssertions();
+            await expect(
+                path.join(tmpPath, 'view/svg', svgPath)).toBeVisuallyEqual(path.join(paths.expectations, '/png/css.vertical.mixed.png')
             );
         });
 
-        it('creates a visually correct stylesheet resource', done => {
+        it('creates a visually correct stylesheet resource', async() => {
+            expect.hasAssertions();
+
             data.css = '../sprite.mixed.css';
 
             const out = mustache.render(previewTemplate, data);
-            const preview = writeFile(path.join(paths.tmp, 'view/html/css.vertical.mixed.html'), out);
-            const previewImage = path.join(paths.tmp, 'view/png/css.vertical.mixed.html.png');
+            const preview = writeFile(path.join(tmpPath, 'view/html/css.vertical.mixed.html'), out);
 
-            preview.should.be.ok;
-
-            capturePuppeteer(preview, previewImage, error => {
-                should(error).not.ok;
-                looksSame(previewImage, path.join(paths.expectations, '/png/css.vertical.mixed.html.png'), (error, result) => {
-                    should(error).not.ok;
-                    should.ok(result.equal, 'The generated CSS preview doesn\'t match the expected one!');
-                    done();
-                });
-            });
+            await expect(preview).toBeVisuallyCorrectAsHTML(
+                path.join(paths.expectations, '/png/css.vertical.mixed.html.png')
+            );
         });
     });
 
@@ -95,9 +80,10 @@ describe(`svg-sprite: with mixed alignment and ${align.length} SVG files`, () =>
         let spriter = null;
         let data = null;
         let svgPath = null;
-        before('creates 2 files', done => {
+
+        beforeAll(async() => {
             spriter = new SVGSpriter({
-                dest: paths.tmp,
+                dest: tmpPath,
                 shape: {
                     align: path.join(paths.fixtures, 'yaml/align.mixed.yaml'),
                     dimension: {
@@ -111,7 +97,7 @@ describe(`svg-sprite: with mixed alignment and ${align.length} SVG files`, () =>
                 }
             });
             addFixtureFiles(spriter, align, cwdAlign);
-            spriter.compile({
+            const { result, data: cssData } = await spriter.compileAsync({
                 view: {
                     sprite: 'svg/view.horizontal.mixed.svg',
                     layout: 'horizontal',
@@ -122,47 +108,32 @@ describe(`svg-sprite: with mixed alignment and ${align.length} SVG files`, () =>
                         }
                     }
                 }
-            }, (error, result, cssData) => {
-                writeFiles(result);
-                data = cssData.view;
-                svgPath = path.basename(result.view.sprite.path);
-                done();
             });
+            writeFiles(result);
+            data = cssData.view;
+            svgPath = path.basename(result.view.sprite.path);
         });
 
-        it('creates visually correct sprite', done => {
-            compareSvg2Png(
-                path.join(paths.tmp, 'view/svg', svgPath),
-                path.join(paths.tmp, 'view/png/css.horizontal.mixed.png'),
-                path.join(paths.expectations, '/png/css.horizontal.mixed.png'),
-                path.join(paths.tmp, 'view/png/css.horizontal.mixed.diff.png'),
-                done,
-                'The horizontal sprite doesn\'t match the expected one!'
+        it('creates visually correct sprite', async() => {
+            expect.hasAssertions();
+            await expect(
+                path.join(tmpPath, 'view/svg', svgPath)).toBeVisuallyEqual(
+                path.join(paths.expectations, '/png/css.horizontal.mixed.png')
             );
         });
 
-        it('creates a visually correct stylesheet resource', done => {
-            sass.render({ file: path.join(paths.tmp, 'view/sprite.mixed.scss') }, (err, scssText) => {
-                should(err).not.ok;
-                should(writeFile(path.join(paths.tmp, 'view/sprite.mixed.scss.css'), scssText.css)).be.ok;
+        it('creates a visually correct stylesheet resource', async() => {
+            expect.hasAssertions();
 
-                data.css = '../sprite.mixed.scss.css';
+            const scssText = sass.renderSync({ file: path.join(tmpPath, 'view/sprite.mixed.scss') });
+            writeFile(path.join(tmpPath, 'view/sprite.mixed.scss.css'), scssText.css);
 
-                const out = mustache.render(previewTemplate, data);
-                const preview = writeFile(path.join(paths.tmp, 'view/html/scss.horizontal.mixed.html'), out);
-                const previewImage = path.join(paths.tmp, 'view/png/scss.horizontal.mixed.html.png');
+            data.css = '../sprite.mixed.scss.css';
 
-                preview.should.be.ok;
+            const out = mustache.render(previewTemplate, data);
+            const preview = writeFile(path.join(tmpPath, 'view/html/scss.horizontal.mixed.html'), out);
 
-                capturePuppeteer(preview, previewImage, error => {
-                    should(error).not.ok;
-                    looksSame(previewImage, path.join(paths.expectations, '/png/css.horizontal.mixed.html.png'), (error, result) => {
-                        should(error).not.ok;
-                        should.ok(result.equal, 'The generated Sass preview doesn\'t match the expected one!');
-                        done();
-                    });
-                });
-            });
+            await expect(preview).toBeVisuallyCorrectAsHTML(path.join(paths.expectations, 'png/css.horizontal.mixed.html.png'));
         });
     });
 });
