+++ test/svg-sprite/minimal-configuration/modes/symbol.test.js
@@ -1,73 +1,59 @@
 'use strict';
 
-/* eslint-disable no-unused-expressions, max-nested-callbacks */
-
 const path = require('path');
 const fs = require('fs');
 const mustache = require('mustache');
-const should = require('should');
-const looksSame = require('looks-same');
 const SVGSpriter = require('../../../../lib/svg-sprite.js');
 const { addFixtureFiles } = require('../../../helpers/add-files.js');
 const writeFiles = require('../../../helpers/write-files.js');
 const writeFile = require('../../../helpers/write-file.js');
-const capturePuppeteer = require('../../../helpers/capture-puppeteer.js');
-const testConfigs = require('../../../helpers/test-configs.js');
+const { constants } = require('../../../helpers/test-configs.js');
 const { paths } = require('../../../helpers/constants.js');
 
 const removeTmpPath = require('../../../helpers/remove-temp-path.js');
 
-testConfigs.forEach(testConfig => {
-    describe(`svg-sprite: ${testConfig.name} in «symbol» mode`, () => {
-        before(removeTmpPath);
+describe.each`
+        name          | testConfigKey
+        ${'default'}  | ${'DEFAULT'}
+        ${'w/o dims'} | ${'WITHOUT_DIMS'}
+`('svg-sprite: $name: «symbol» mode', ({ testConfigKey }) => {
+    const testConfig = constants[testConfigKey];
 
-        let svg;
-        let data = {};
-        let spriter;
-        before('creates 2 files for packed layout', done => {
-            spriter = new SVGSpriter({
-                dest: paths.tmp
-            });
-            addFixtureFiles(spriter, testConfig.files, testConfig.cwd);
-            spriter.compile({
-                symbol: {
-                    sprite: `svg/symbol${testConfig.namespace}.svg`,
-                    render: {
-                        css: true
-                    }
-                }
-            }, (error, result, cssData) => {
-                writeFiles(result).should.be.exactly(2);
-                data = cssData.symbol;
-                svg = path.basename(result.symbol.sprite.path);
-                done();
-            });
+    const tmpPath = path.join(paths.tmp, `symbol${testConfig.namespace}`);
+
+    let svg;
+    let spriter;
+    let data;
+
+    beforeAll(async() => {
+        await removeTmpPath(tmpPath);
+        data = {};
+
+        spriter = new SVGSpriter({
+            dest: tmpPath
         });
-
-        it('creates a visually correct stylesheet resource in CSS format', done => {
-            data.svg = fs.readFileSync(path.join(paths.tmp, 'symbol/svg', svg)).toString();
-            data.css = '../sprite.css';
-            const previewTemplate = fs.readFileSync(path.join(__dirname, '../../../tmpl/symbol.html'), 'utf-8');
-            const out = mustache.render(previewTemplate, data);
-            const preview = writeFile(path.join(paths.tmp, 'symbol/html/symbol.html'), out);
-            const previewImage = path.join(paths.tmp, `symbol/symbol.html${testConfig.namespace}.png`);
-            preview.should.be.ok;
-
-            capturePuppeteer(preview, previewImage, error => {
-                should(error).not.ok;
-                if (error) {
-                    return done(error);
+        addFixtureFiles(spriter, testConfig.files, testConfig.cwd);
+        const { result, data: cssData } = await spriter.compileAsync({
+            symbol: {
+                sprite: `svg/symbol${testConfig.namespace}.svg`, render: {
+                    css: true
                 }
-
-                looksSame(
-                    previewImage,
-                    path.join(paths.expectations, `png/symbol.html${testConfig.namespace}.png`),
-                    (error, result) => {
-                        should(error).not.ok;
-                        should.ok(result.equal, 'The generated CSS preview doesn\'t match the expected one!');
-                        done();
-                    });
-            });
+            }
         });
+        writeFiles(result);
+        data = cssData.symbol;
+        svg = path.basename(result.symbol.sprite.path);
+    });
+
+    it('creates a visually correct stylesheet resource in CSS format', async() => {
+        expect.hasAssertions();
+
+        data.svg = fs.readFileSync(path.join(tmpPath, 'symbol/svg', svg)).toString();
+        data.css = '../sprite.css';
+        const previewTemplate = fs.readFileSync(path.join(__dirname, '../../../tmpl/symbol.html'), 'utf-8');
+        const out = mustache.render(previewTemplate, data);
+        const preview = writeFile(path.join(tmpPath, 'symbol/html/symbol.html'), out);
+
+        await expect(preview).toBeVisuallyCorrectAsHTML(path.join(paths.expectations, `png/symbol.html${testConfig.namespace}.png`));
     });
 });
