+++ test/svg-sprite/minimal-configuration/modes/view.packed.test.js
@@ -1,35 +1,40 @@
 'use strict';
 
-/* eslint-disable no-unused-expressions, max-nested-callbacks */
 const path = require('path');
 const fs = require('fs');
 const mustache = require('mustache');
-const should = require('should');
-const looksSame = require('looks-same');
 const writeFiles = require('../../../helpers/write-files.js');
-const compareSvg2Png = require('../../../helpers/compare-svg-2-png.js');
 const writeFile = require('../../../helpers/write-file.js');
-const capturePuppeteer = require('../../../helpers/capture-puppeteer.js');
-const testConfigs = require('../../../helpers/test-configs.js');
 const SVGSpriter = require('../../../../lib/svg-sprite.js');
 const { addFixtureFiles } = require('../../../helpers/add-files.js');
 const { paths } = require('../../../helpers/constants.js');
 
 const removeTmpPath = require('../../../helpers/remove-temp-path.js');
+const { constants } = require('../../../helpers/test-configs.js');
 
-testConfigs.forEach(testConfig => {
+describe.each`
+        name          | testConfigKey
+        ${'default'}  | ${'DEFAULT'}
+        ${'w/o dims'} | ${'WITHOUT_DIMS'}
+`('svg-sprite: $name: «symbol» mode', ({ testConfigKey }) => {
+    const testConfig = constants[testConfigKey];
+
+    const tmpPath = path.join(paths.tmp, `view${testConfig.namespace}.packed`);
     let spriter;
     let svg;
-    let data = {};
+    let data;
+
+    beforeAll(removeTmpPath.bind(null, tmpPath));
+
     // Test the view mode
     describe(`svg-sprite: ${testConfig.name} in «view» mode`, () => {
-        before(removeTmpPath);
-        before('creates 2 files for packed layout', done => {
+        beforeAll(async() => {
+            data = {};
             spriter = new SVGSpriter({
-                dest: paths.tmp
+                dest: tmpPath
             });
             addFixtureFiles(spriter, testConfig.files, testConfig.cwd);
-            spriter.compile({
+            const { result, data: cssData } = await spriter.compileAsync({
                 view: {
                     sprite: `svg/view.packed${testConfig.namespace}.svg`,
                     layout: 'packed',
@@ -38,45 +43,30 @@ testConfigs.forEach(testConfig => {
                         css: true
                     }
                 }
-            }, (error, result, cssData) => {
-                writeFiles(result);
-                data = cssData.view;
-                svg = path.basename(result.view.sprite.path);
-                done();
             });
+            writeFiles(result);
+            data = cssData.view;
+            svg = path.basename(result.view.sprite.path);
         });
 
         // Packed layout
-        it('creates visually correct sprite with packed layout', done => {
-            compareSvg2Png(
-                path.join(paths.tmp, 'view/svg', svg),
-                path.join(paths.tmp, `view/png/view.packed${testConfig.namespace}.png`),
-                path.join(paths.expectations, `png/css.packed${testConfig.namespace}.png`),
-                path.join(paths.tmp, `view/png/view.packed${testConfig.namespace}.diff.png`),
-                done,
-                'The packed sprite doesn\'t match the expected one!'
+        it('creates visually correct sprite with packed layout', async() => {
+            expect.hasAssertions();
+            await expect(
+                path.join(tmpPath, 'view/svg', svg)).toBeVisuallyEqual(
+                path.join(paths.expectations, `png/css.packed${testConfig.namespace}.png`)
             );
         });
 
-        it('creates a visually correct stylesheet resource in CSS format', done => {
+        it('creates a visually correct stylesheet resource in CSS format', async() => {
+            expect.hasAssertions();
+
             data.css = '../sprite.css';
             const previewTemplate = fs.readFileSync(path.join(__dirname, '../../../tmpl/view.html'), 'utf-8');
             const out = mustache.render(previewTemplate, data);
-            const preview = writeFile(path.join(paths.tmp, 'view/html/view.html'), out);
-            const previewImage = path.join(paths.tmp, `view/png/view.html${testConfig.namespace}.png`);
-            preview.should.be.ok;
+            const preview = writeFile(path.join(tmpPath, 'view/html/view.html'), out);
 
-            capturePuppeteer(preview, previewImage, error => {
-                should(error).not.ok;
-                looksSame(
-                    previewImage,
-                    path.join(paths.expectations, `png/view.html${testConfig.namespace}.png`),
-                    (error, result) => {
-                        should(error).not.ok;
-                        should.ok(result.equal, 'The generated CSS preview doesn\'t match the expected one!');
-                        done();
-                    });
-            });
+            await expect(preview).toBeVisuallyCorrectAsHTML(path.join(paths.expectations, `png/view.html${testConfig.namespace}.png`));
         });
     });
 });
